requirements.txt
absl-py>=0.5.0
numpy>=1.12.1
tensorflow>=1.11.0
